techno_farm/requirements.txt
djangorestframework==3.14.0
PyJWT==1.7.1